+++ protected/js/datepicker.js
@@ -1,3 +1,7 @@
 $(document).ready(function() {
-  $('[data-js="datepicker"]').datepicker();
+  $('[data-js="datepicker"]').datepicker({
+    dateFormat: 'yy-mm-dd',
+    changeMonth: true,
+    changeYear: true
+  });
 })